requirements.txt
requests==2.31.0
cryptography==41.0.5
matplotlib==3.8.2
python-dateutil==2.8.2
Flask==3.0.0
python-dotenv==1.0.0
pandas==2.1.3
scikit-learn==1.3.2
scipy==1.11.4
threadpoolctl==3.2.0
inflection==0.5.1
openapi==1.1.0
jsonify==0.5
tensorflow==2.14.0
numpy==1.23.5
gunicorn==21.2.0